Code cleanup and style
- Remove unnecessary exceptions
- Inline redundant variables
- Make members final
- Updates to method visibilities
- etc.

+++ src/main/java/io/github/spannm/jackcess/impl/query/QueryImpl.java
@@ -453,7 +453,7 @@ public abstract class QueryImpl implements Query {
     }
 
     protected static int getIntValue(Integer i, int def) {
-        return i != null ? (int) i : def;
+        return i != null ? i : def;
     }
 
     protected static StringBuilder toOptionalQuotedExpr(StringBuilder builder,
